pyproject.toml
[project]
name = "dionysus-core"
version = "0.1.0"
description = "Lightweight consciousness system for Claude MCP integration"
authors = [{name = "Mani Saint Victor"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.2",

    # AI/ML
    "openai>=1.3.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.1",

    # MCP Server
    "mcp>=1.0.0",

    # Knowledge Graph (Neo4j via Graphiti)
    "neo4j>=5.14.0",
    "graphiti-core>=0.24.3",

    # Agents
    "smolagents>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pytest-mock>=3.11.1",
    "pytest-cov>=4.1.0",
]

[project.scripts]
dionysus-core = "dionysus_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "dionysus_mcp*", "mcp*", "scripts*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "."]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
